feat: add `sql.query` function
Add `sql.query` function, similar to the Neon driver.

The difference from the `sql` template literal is that the user is
expected to provide a raw query with placeholders and a separate array
of values.

The difference from `Client#query`, which accepts the same arguments, is
that `sql.query` returns deserialized results as an array of objects and
not the raw response.